+++ src/components/Calendar/Calendar.types.ts
@@ -3,3 +3,12 @@ export interface CalenderProps {
   value: Date;
   setValue: (button: 'prev' | 'next') => void;
 }
+
+export interface TaskData {
+  id: number;
+  title: string;
+  startDate: Date;
+  endDate: Date;
+  status: number;
+  description: string;
+}
